refactor: updated client names for better clarity
