+++ src/weather-client/weather-client.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
-import { HttpWeatherClient } from './http-weather-client';
+import { WeatherApiClient } from './weather-api-client';
 import { HttpModule } from '../infrastructure/http/http.module';
 import { CacheModule } from '../cache/cache.module';
-import { OpenWeatherClient } from './openweather-client';
+import { OpenWeatherClient } from './open-weather-client';
 
 @Module({
   imports: [HttpModule, CacheModule],
-  providers: [OpenWeatherClient, HttpWeatherClient],
-  exports: [HttpWeatherClient, OpenWeatherClient],
+  providers: [OpenWeatherClient, WeatherApiClient],
+  exports: [WeatherApiClient, OpenWeatherClient],
 })
 export class WeatherClientModule {}
